+++ core/src/include/jcanvas/core/jindexedimage.h
@@ -82,7 +82,7 @@ class IndexedImage : public Image {
      * \brief
      *
      */
-    virtual std::shared_ptr<Image> Rotate(double radians, bool resize = true);
+    virtual std::shared_ptr<Image> Rotate(float radians, bool resize = true);
     
     /**
      * \brief
@@ -100,7 +100,7 @@ class IndexedImage : public Image {
      * \brief
      *
      */
-    virtual std::shared_ptr<Image> Blend(double alpha);
+    virtual std::shared_ptr<Image> Blend(float alpha);
 
     /**
      * \brief
